+++ viforest/app/layout.tsx
@@ -20,8 +20,9 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
         <Providers>
           <DragProvider>
             <div className="flex flex-col min-h-screen">
-              <main className="flex-1">{children}</main>
-              <Navbar />
+              <main className="flex-1">
+                {children}
+              </main>
             </div>
           </DragProvider>
         </Providers>