+++ viforest/src-tauri/src/lib.rs
@@ -1,13 +1,14 @@
-use tauri::Manager;
+//use tauri::Manager;
 #[cfg_attr(mobile, tauri::mobile_entry_point)]
 pub fn run() {
     tauri::Builder::default()
+        .plugin(tauri_plugin_opener::init())
         .plugin(tauri_plugin_fs::init())
         .plugin(tauri_plugin_http::init())
         .setup(|app| {
             if cfg!(debug_assertions) {
-                let window = app.get_webview_window("main").unwrap();
-                window.open_devtools();
+                //let window = app.get_webview_window("main").unwrap();
+                //window.open_devtools();
                 app.handle().plugin(
                     tauri_plugin_log::Builder::default()
                         .level(log::LevelFilter::Info)