+++ viforest/CHANGELOG.md
@@ -1,15 +1,19 @@
 # High Priority:
 - [x] Fix downloading with Tauri's `fs` plugin
-- [ ] Fix weird EOF error on first startup
-- [ ] Fix navbar blocking explorer
+- [x] Fix navbar blocking explorer
+- [x] Fix dot pattern being blocked by title
+- [x] Fix file explorer not populating on connection
+- [x] Fix weird EOF error on first startup
+- [x] Support revealing files on download & opening links in modals
 
 # Low Priority:
 - [ ] Fix logo (it's so goofy)
 - [ ] Fix the drag and drop animation
-- [ ] Fix UI not supporting different screen sizes
+- [ ] Fix UI not supporting different small screens
 - [ ] Fix device connection flash when clicking into the dialog
 
 # Planned Features:
 - [ ] Full file tree traversal on initial load to eliminate loading state when clicking between folders
 - [ ] Option to choose file format other than PDF for download
 - [ ] Support for pasting pages directly from OneNote and auto-converting to PDF
+- [ ] Native file explorer integration (right click to send)